Protected updating and deleting a product

+++ controllers/feed.js
@@ -126,6 +126,11 @@ exports.updatePost = (req, res, next) => {
 				error.statusCode = 404;
 				throw error;
 			}
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error('Not authorized!');
+				error.statusCode = 403;
+				throw error;
+			}
 			if (imageUrl !== post.imageUrl) {
 				clearImage(post.imageUrl);
 			}
@@ -155,11 +160,15 @@ exports.deletePost = (req, res, next) => {
 				throw error;
 			}
 			//Check if the loggedIn user created this post
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error('Not authorized!');
+				error.statusCode = 403;
+				throw error;
+			}
 			clearImage(post.imageUrl);
 			return Post.findByIdAndRemove(postId);
 		})
 		.then((result) => {
-			console.log(result);
 			res.status(200).json({ message: 'Deleted post.' });
 		})
 		.catch((err) => {
